Clean: consolidate functions
