+++ crates/rbuilder/src/backtest/mod.rs
@@ -1,6 +1,5 @@
 mod backtest_build_block;
 mod backtest_build_range;
-pub mod config; 
 pub mod execute;
 pub mod fetch;
